+++ files_sdk/models/automation.py
@@ -19,6 +19,7 @@ class Automation:
         'user_id': None,     # int64 - User ID of the Automation's creator.
         'user_ids': None,     # array - IDs of Users for the Automation (i.e. who to Request File from)
         'group_ids': None,     # array - IDs of Groups for the Automation (i.e. who to Request File from)
+        'webhook_url': None,     # string - If trigger is `webhook`, this is the URL of the webhook to trigger the Automation.
     }
 
     def __init__(self, attributes=None, options=None):
@@ -113,7 +114,7 @@ class Automation:
 # Parameters:
 #   cursor - string - Used for pagination.  Send a cursor value to resume an existing list from the point at which you left off.  Get a cursor from an existing list via the X-Files-Cursor-Next header.
 #   per_page - int64 - Number of records to show per page.  (Max: 10,000, 1,000 or less is recommended).
-#   sort_by - object - If set, sort records by the specified field in either 'asc' or 'desc' direction (e.g. sort_by[last_login_at]=desc). Valid fields are `site_id` and `automation`.
+#   sort_by - object - If set, sort records by the specified field in either 'asc' or 'desc' direction (e.g. sort_by[last_login_at]=desc). Valid fields are `automation`.
 #   filter - object - If set, return records where the specifiied field is equal to the supplied value. Valid fields are `automation`.
 #   filter_gt - object - If set, return records where the specifiied field is greater than the supplied value. Valid fields are `automation`.
 #   filter_gteq - object - If set, return records where the specifiied field is greater than or equal to the supplied value. Valid fields are `automation`.