+++ files_sdk/models/bandwidth_snapshot.py
@@ -8,6 +8,8 @@ class BandwidthSnapshot:
         'id': None,     # int64 - Site bandwidth ID
         'bytes_received': None,     # double - Site bandwidth report bytes received
         'bytes_sent': None,     # double - Site bandwidth report bytes sent
+        'sync_bytes_received': None,     # double - Site sync bandwidth report bytes received
+        'sync_bytes_sent': None,     # double - Site sync bandwidth report bytes sent
         'requests_get': None,     # double - Site bandwidth report get requests
         'requests_put': None,     # double - Site bandwidth report put requests
         'requests_other': None,     # double - Site bandwidth report other requests
@@ -35,7 +37,7 @@ class BandwidthSnapshot:
 # Parameters:
 #   cursor - string - Used for pagination.  Send a cursor value to resume an existing list from the point at which you left off.  Get a cursor from an existing list via the X-Files-Cursor-Next header.
 #   per_page - int64 - Number of records to show per page.  (Max: 10,000, 1,000 or less is recommended).
-#   sort_by - object - If set, sort records by the specified field in either 'asc' or 'desc' direction (e.g. sort_by[last_login_at]=desc). Valid fields are `site_id` and `logged_at`.
+#   sort_by - object - If set, sort records by the specified field in either 'asc' or 'desc' direction (e.g. sort_by[last_login_at]=desc). Valid fields are `logged_at`.
 #   filter - object - If set, return records where the specifiied field is equal to the supplied value. Valid fields are `logged_at`.
 #   filter_gt - object - If set, return records where the specifiied field is greater than the supplied value. Valid fields are `logged_at`.
 #   filter_gteq - object - If set, return records where the specifiied field is greater than or equal to the supplied value. Valid fields are `logged_at`.