+++ tests/test_inbox_registration.py
@@ -2,16 +2,13 @@ import unittest
 import inspect
 import files_sdk
 from tests.base import TestBase
-from files_sdk.models import BundleDownload
-from files_sdk import bundle_download
+from files_sdk.models import InboxRegistration
+from files_sdk import inbox_registration
 
-class BundleDownloadTest(TestBase):
+class InboxRegistrationTest(TestBase):
     pass 
     # Instance Methods
 
     # Static Methods
-    def test_list(self):
-        resp = bundle_download.list()
-
 if __name__ == '__main__':
     unittest.main()