+++ tests/test_inbox_upload.py
@@ -2,16 +2,16 @@ import unittest
 import inspect
 import files_sdk
 from tests.base import TestBase
-from files_sdk.models import BundleDownload
-from files_sdk import bundle_download
+from files_sdk.models import InboxUpload
+from files_sdk import inbox_upload
 
-class BundleDownloadTest(TestBase):
+class InboxUploadTest(TestBase):
     pass 
     # Instance Methods
 
     # Static Methods
     def test_list(self):
-        resp = bundle_download.list()
+        resp = inbox_upload.list()
 
 if __name__ == '__main__':
     unittest.main()